# feat: optimize ABI generation time