--- a/cargo-near/src/abi/generation.rs
+++ b/cargo-near/src/abi/generation.rs
@@ -25,9 +25,10 @@ pub(crate) fn generate_toml(manifest_path: &CargoManifestPath) -> anyhow::Result
         .as_table_mut()
         .expect("[dependencies] is a table specified in the template");
 
-    // Make near-sdk dependency use default features
+    // Make near-sdk dependency not use default features to save on compilation time, but ensure `abi` is enabled
     near_sdk.remove("default-features");
     near_sdk.remove("optional");
+    near_sdk.insert("default-features".to_string(), value::Value::Boolean(false));
     near_sdk.insert(
         "features".to_string(),
         value::Value::Array(vec![value::Value::String("abi".to_string())]),