--- a/cargo-near/templates/_Cargo.toml
+++ b/cargo-near/templates/_Cargo.toml
@@ -9,6 +9,11 @@ build = "build.rs"
 name = "near-abi-gen"
 path = "main.rs"
 
+[profile.dev]
+opt-level = 0
+debug = 0
+lto = "off"
+
 [dependencies]
 serde_json = "1.0"