Fixes #8 

This PR introduces three compilation time improvements:
1. Removes `nearcore` crate dependencies from the generated abigen crates (reduces compilation time from 22.5s to 10.5s on `abi` example in NEAR SDK).
2. Makes the generated abigen crates use `dev` profile with disabled optimizations and disabled LTO (reduces compilation town from 10.5s further to 9.5s)
3. Makes the original (user-written) crate use `dev` profile with disabled optimizations and disabled LTO through cargo envvars (reduces dylib compilation time from 25s to 21.5s).